LRCI-3760 Rename all EntityRelationshipDALO classes
