+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/bui1d/dalo/BuildToBuildRunsEntityRelationshipDALO.java
@@ -7,8 +7,8 @@ package com.liferay.jethr0.bui1d.dalo;
 
 import com.liferay.jethr0.bui1d.Build;
 import com.liferay.jethr0.bui1d.BuildFactory;
-import com.liferay.jethr0.bui1d.parameter.BuildParameter;
-import com.liferay.jethr0.bui1d.parameter.BuildParameterFactory;
+import com.liferay.jethr0.bui1d.run.BuildRun;
+import com.liferay.jethr0.bui1d.run.BuildRunFactory;
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
 
@@ -19,12 +19,12 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class BuildToBuildParametersDALO
-	extends BaseEntityRelationshipDALO<Build, BuildParameter> {
+public class BuildToBuildRunsEntityRelationshipDALO
+	extends BaseEntityRelationshipDALO<Build, BuildRun> {
 
 	@Override
-	public EntityFactory<BuildParameter> getChildEntityFactory() {
-		return _buildParameterFactory;
+	public EntityFactory<BuildRun> getChildEntityFactory() {
+		return _buildRunFactory;
 	}
 
 	@Override
@@ -34,13 +34,13 @@ public class BuildToBuildParametersDALO
 
 	@Override
 	protected String getObjectRelationshipName() {
-		return "buildToBuildParameters";
+		return "buildToBuildRuns";
 	}
 
 	@Autowired
 	private BuildFactory _buildFactory;
 
 	@Autowired
-	private BuildParameterFactory _buildParameterFactory;
+	private BuildRunFactory _buildRunFactory;
 
 }
