+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/bui1d/dalo/BuildToEnvironmentsEntityRelationshipDALO.java
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.project.dalo;
+package com.liferay.jethr0.bui1d.dalo;
 
 import com.liferay.jethr0.bui1d.Build;
 import com.liferay.jethr0.bui1d.BuildFactory;
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
-import com.liferay.jethr0.project.Project;
-import com.liferay.jethr0.project.ProjectFactory;
+import com.liferay.jethr0.environment.Environment;
+import com.liferay.jethr0.environment.EnvironmentFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
@@ -19,28 +19,28 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class ProjectToBuildsDALO
-	extends BaseEntityRelationshipDALO<Project, Build> {
+public class BuildToEnvironmentsEntityRelationshipDALO
+	extends BaseEntityRelationshipDALO<Build, Environment> {
 
 	@Override
-	public EntityFactory<Build> getChildEntityFactory() {
+	public EntityFactory<Environment> getChildEntityFactory() {
+		return _environmentFactory;
+	}
+
+	@Override
+	public EntityFactory<Build> getParentEntityFactory() {
 		return _buildFactory;
 	}
 
 	@Override
-	public EntityFactory<Project> getParentEntityFactory() {
-		return _projectFactory;
-	}
-
-	@Override
 	protected String getObjectRelationshipName() {
-		return "projectToBuilds";
+		return "buildToEnvironments";
 	}
 
 	@Autowired
 	private BuildFactory _buildFactory;
 
 	@Autowired
-	private ProjectFactory _projectFactory;
+	private EnvironmentFactory _environmentFactory;
 
 }
