+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/environment/repository/EnvironmentRepository.java
@@ -6,7 +6,7 @@
 package com.liferay.jethr0.environment.repository;
 
 import com.liferay.jethr0.bui1d.Build;
-import com.liferay.jethr0.bui1d.dalo.BuildToEnvironmentsDALO;
+import com.liferay.jethr0.bui1d.dalo.BuildToEnvironmentsEntityRelationshipDALO;
 import com.liferay.jethr0.entity.repository.BaseEntityRepository;
 import com.liferay.jethr0.environment.Environment;
 import com.liferay.jethr0.environment.dalo.EnvironmentDALO;
@@ -26,8 +26,8 @@ public class EnvironmentRepository extends BaseEntityRepository<Environment> {
 	public Set<Environment> getAll(Build build) {
 		Set<Environment> buildEnvironments = new HashSet<>();
 
-		Set<Long> environmentIds = _buildToEnvironmentsDALO.getChildEntityIds(
-			build);
+		Set<Long> environmentIds =
+			_buildToEnvironmentsEntityRelationshipDALO.getChildEntityIds(build);
 
 		for (Environment environment : getAll()) {
 			if (!environmentIds.contains(environment.getId())) {
@@ -50,7 +50,8 @@ public class EnvironmentRepository extends BaseEntityRepository<Environment> {
 	}
 
 	@Autowired
-	private BuildToEnvironmentsDALO _buildToEnvironmentsDALO;
+	private BuildToEnvironmentsEntityRelationshipDALO
+		_buildToEnvironmentsEntityRelationshipDALO;
 
 	@Autowired
 	private EnvironmentDALO _environmentDALO;
