+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/gitbranch/repository/GitBranchRepository.java
@@ -9,7 +9,7 @@ import com.liferay.jethr0.entity.repository.BaseEntityRepository;
 import com.liferay.jethr0.gitbranch.GitBranch;
 import com.liferay.jethr0.gitbranch.dalo.GitBranchDALO;
 import com.liferay.jethr0.project.Project;
-import com.liferay.jethr0.project.dalo.ProjectsToGitBranchesDALO;
+import com.liferay.jethr0.project.dalo.ProjectsToGitBranchesEntityRelationshipDALO;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -26,8 +26,9 @@ public class GitBranchRepository extends BaseEntityRepository<GitBranch> {
 	public Set<GitBranch> getAll(Project project) {
 		Set<GitBranch> projectGitBranches = new HashSet<>();
 
-		Set<Long> gitBranchIds = _projectsToGitBranchesDALO.getChildEntityIds(
-			project);
+		Set<Long> gitBranchIds =
+			_projectsToGitBranchesEntityRelationshipDALO.getChildEntityIds(
+				project);
 
 		for (GitBranch gitBranch : getAll()) {
 			if (!gitBranchIds.contains(gitBranch.getId())) {
@@ -53,6 +54,7 @@ public class GitBranchRepository extends BaseEntityRepository<GitBranch> {
 	private GitBranchDALO _gitBranchDALO;
 
 	@Autowired
-	private ProjectsToGitBranchesDALO _projectsToGitBranchesDALO;
+	private ProjectsToGitBranchesEntityRelationshipDALO
+		_projectsToGitBranchesEntityRelationshipDALO;
 
 }
