+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/jenkins/dalo/JenkinsCohortToJenkinsServersEntityRelationshipDALO.java
@@ -1,5 +1,5 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
@@ -7,8 +7,8 @@ package com.liferay.jethr0.jenkins.dalo;
 
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
-import com.liferay.jethr0.jenkins.node.JenkinsNode;
-import com.liferay.jethr0.jenkins.node.JenkinsNodeFactory;
+import com.liferay.jethr0.jenkins.cohort.JenkinsCohort;
+import com.liferay.jethr0.jenkins.cohort.JenkinsCohortFactory;
 import com.liferay.jethr0.jenkins.server.JenkinsServer;
 import com.liferay.jethr0.jenkins.server.JenkinsServerFactory;
 
@@ -19,26 +19,26 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class JenkinsServerToJenkinsNodesDALO
-	extends BaseEntityRelationshipDALO<JenkinsServer, JenkinsNode> {
+public class JenkinsCohortToJenkinsServersEntityRelationshipDALO
+	extends BaseEntityRelationshipDALO<JenkinsCohort, JenkinsServer> {
 
 	@Override
-	public EntityFactory<JenkinsNode> getChildEntityFactory() {
-		return _jenkinsNodeFactory;
-	}
-
-	@Override
-	public EntityFactory<JenkinsServer> getParentEntityFactory() {
+	public EntityFactory<JenkinsServer> getChildEntityFactory() {
 		return _jenkinsServerFactory;
 	}
 
 	@Override
+	public EntityFactory<JenkinsCohort> getParentEntityFactory() {
+		return _jenkinsCohortFactory;
+	}
+
+	@Override
 	protected String getObjectRelationshipName() {
-		return "jenkinsServerToJenkinsNodes";
+		return "jenkinsCohortToJenkinsServers";
 	}
 
 	@Autowired
-	private JenkinsNodeFactory _jenkinsNodeFactory;
+	private JenkinsCohortFactory _jenkinsCohortFactory;
 
 	@Autowired
 	private JenkinsServerFactory _jenkinsServerFactory;
