+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/jenkins/repository/JenkinsCohortRepository.java
@@ -8,7 +8,7 @@ package com.liferay.jethr0.jenkins.repository;
 import com.liferay.jethr0.entity.repository.BaseEntityRepository;
 import com.liferay.jethr0.jenkins.cohort.JenkinsCohort;
 import com.liferay.jethr0.jenkins.dalo.JenkinsCohortDALO;
-import com.liferay.jethr0.jenkins.dalo.JenkinsCohortToJenkinsServersDALO;
+import com.liferay.jethr0.jenkins.dalo.JenkinsCohortToJenkinsServersEntityRelationshipDALO;
 import com.liferay.jethr0.util.StringUtil;
 
 import java.util.Objects;
@@ -58,8 +58,8 @@ public class JenkinsCohortRepository
 	public void initializeRelationships() {
 		for (JenkinsCohort jenkinsCohort : getAll()) {
 			for (long jenkinsServerId :
-					_jenkinsCohortToJenkinsServersDALO.getChildEntityIds(
-						jenkinsCohort)) {
+					_jenkinsCohortToJenkinsServersEntityRelationshipDALO.
+						getChildEntityIds(jenkinsCohort)) {
 
 				if (jenkinsServerId == 0) {
 					continue;
@@ -81,8 +81,8 @@ public class JenkinsCohortRepository
 	private JenkinsCohortDALO _jenkinsCohortDALO;
 
 	@Autowired
-	private JenkinsCohortToJenkinsServersDALO
-		_jenkinsCohortToJenkinsServersDALO;
+	private JenkinsCohortToJenkinsServersEntityRelationshipDALO
+		_jenkinsCohortToJenkinsServersEntityRelationshipDALO;
 
 	private JenkinsServerRepository _jenkinsServerRepository;
