+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/jenkins/repository/JenkinsServerRepository.java
@@ -8,7 +8,7 @@ package com.liferay.jethr0.jenkins.repository;
 import com.liferay.jethr0.entity.repository.BaseEntityRepository;
 import com.liferay.jethr0.jenkins.cohort.JenkinsCohort;
 import com.liferay.jethr0.jenkins.dalo.JenkinsServerDALO;
-import com.liferay.jethr0.jenkins.dalo.JenkinsServerToJenkinsNodesDALO;
+import com.liferay.jethr0.jenkins.dalo.JenkinsServerToJenkinsNodesEntityRelationshipDALO;
 import com.liferay.jethr0.jenkins.server.JenkinsServer;
 import com.liferay.jethr0.util.StringUtil;
 
@@ -131,8 +131,8 @@ public class JenkinsServerRepository
 			jenkinsServer.setJenkinsCohort(jenkinsCohort);
 
 			for (long jenkinsNodeId :
-					_jenkinsServerToJenkinsNodesDALO.getChildEntityIds(
-						jenkinsServer)) {
+					_jenkinsServerToJenkinsNodesEntityRelationshipDALO.
+						getChildEntityIds(jenkinsServer)) {
 
 				if (jenkinsNodeId == 0) {
 					continue;
@@ -169,7 +169,8 @@ public class JenkinsServerRepository
 	private JenkinsServerDALO _jenkinsServerDALO;
 
 	@Autowired
-	private JenkinsServerToJenkinsNodesDALO _jenkinsServerToJenkinsNodesDALO;
+	private JenkinsServerToJenkinsNodesEntityRelationshipDALO
+		_jenkinsServerToJenkinsNodesEntityRelationshipDALO;
 
 	@Value("${jenkins.user.name}")
 	private String _jenkinsUserName;
