+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/project/dalo/ProjectPrioritizerToProjectComparatorsEntityRelationshipDALO.java
@@ -19,7 +19,7 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class ProjectPrioritizerToProjectComparatorsDALO
+public class ProjectPrioritizerToProjectComparatorsEntityRelationshipDALO
 	extends BaseEntityRelationshipDALO<ProjectPrioritizer, ProjectComparator> {
 
 	@Override
