+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/project/dalo/ProjectToTasksEntityRelationshipDALO.java
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.task.dalo;
+package com.liferay.jethr0.project.dalo;
 
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
-import com.liferay.jethr0.environment.Environment;
-import com.liferay.jethr0.environment.EnvironmentFactory;
+import com.liferay.jethr0.project.Project;
+import com.liferay.jethr0.project.ProjectFactory;
 import com.liferay.jethr0.task.Task;
 import com.liferay.jethr0.task.TaskFactory;
 
@@ -19,26 +19,26 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class TaskToEnvironmentsDALO
-	extends BaseEntityRelationshipDALO<Task, Environment> {
+public class ProjectToTasksEntityRelationshipDALO
+	extends BaseEntityRelationshipDALO<Project, Task> {
 
 	@Override
-	public EntityFactory<Environment> getChildEntityFactory() {
-		return _environmentFactory;
-	}
-
-	@Override
-	public EntityFactory<Task> getParentEntityFactory() {
+	public EntityFactory<Task> getChildEntityFactory() {
 		return _taskFactory;
 	}
 
 	@Override
+	public EntityFactory<Project> getParentEntityFactory() {
+		return _projectFactory;
+	}
+
+	@Override
 	protected String getObjectRelationshipName() {
-		return "taskToEnvironments";
+		return "projectToTasks";
 	}
 
 	@Autowired
-	private EnvironmentFactory _environmentFactory;
+	private ProjectFactory _projectFactory;
 
 	@Autowired
 	private TaskFactory _taskFactory;
