+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/project/dalo/ProjectsToGitBranchesEntityRelationshipDALO.java
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.testsuite.dalo;
+package com.liferay.jethr0.project.dalo;
 
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
+import com.liferay.jethr0.gitbranch.GitBranch;
+import com.liferay.jethr0.gitbranch.GitBranchFactory;
 import com.liferay.jethr0.project.Project;
 import com.liferay.jethr0.project.ProjectFactory;
-import com.liferay.jethr0.testsuite.TestSuite;
-import com.liferay.jethr0.testsuite.TestSuiteFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
@@ -19,12 +19,12 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class ProjectsToTestSuitesDALO
-	extends BaseEntityRelationshipDALO<Project, TestSuite> {
+public class ProjectsToGitBranchesEntityRelationshipDALO
+	extends BaseEntityRelationshipDALO<Project, GitBranch> {
 
 	@Override
-	public EntityFactory<TestSuite> getChildEntityFactory() {
-		return _testSuiteFactory;
+	public EntityFactory<GitBranch> getChildEntityFactory() {
+		return _gitBranchFactory;
 	}
 
 	@Override
@@ -34,13 +34,13 @@ public class ProjectsToTestSuitesDALO
 
 	@Override
 	protected String getObjectRelationshipName() {
-		return "projectsToTestSuites";
+		return "projectsToGitBranches";
 	}
 
 	@Autowired
-	private ProjectFactory _projectFactory;
+	private GitBranchFactory _gitBranchFactory;
 
 	@Autowired
-	private TestSuiteFactory _testSuiteFactory;
+	private ProjectFactory _projectFactory;
 
 }
