+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/project/dalo/ProjectsToJenkinsCohortsEntityRelationshipDALO.java
@@ -19,7 +19,7 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class ProjectsToJenkinsCohortsDALO
+public class ProjectsToJenkinsCohortsEntityRelationshipDALO
 	extends BaseEntityRelationshipDALO<Project, JenkinsCohort> {
 
 	@Override
