+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/testsuite/dalo/ProjectsToTestSuitesEntityRelationshipDALO.java
@@ -3,14 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.project.dalo;
+package com.liferay.jethr0.testsuite.dalo;
 
 import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
 import com.liferay.jethr0.entity.factory.EntityFactory;
 import com.liferay.jethr0.project.Project;
 import com.liferay.jethr0.project.ProjectFactory;
-import com.liferay.jethr0.task.Task;
-import com.liferay.jethr0.task.TaskFactory;
+import com.liferay.jethr0.testsuite.TestSuite;
+import com.liferay.jethr0.testsuite.TestSuiteFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
@@ -19,12 +19,12 @@ import org.springframework.context.annotation.Configuration;
  * @author Michael Hashimoto
  */
 @Configuration
-public class ProjectToTasksDALO
-	extends BaseEntityRelationshipDALO<Project, Task> {
+public class ProjectsToTestSuitesEntityRelationshipDALO
+	extends BaseEntityRelationshipDALO<Project, TestSuite> {
 
 	@Override
-	public EntityFactory<Task> getChildEntityFactory() {
-		return _taskFactory;
+	public EntityFactory<TestSuite> getChildEntityFactory() {
+		return _testSuiteFactory;
 	}
 
 	@Override
@@ -34,13 +34,13 @@ public class ProjectToTasksDALO
 
 	@Override
 	protected String getObjectRelationshipName() {
-		return "projectToTasks";
+		return "projectsToTestSuites";
 	}
 
 	@Autowired
 	private ProjectFactory _projectFactory;
 
 	@Autowired
-	private TaskFactory _taskFactory;
+	private TestSuiteFactory _testSuiteFactory;
 
 }
