+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/testsuite/repository/TestSuiteRepository.java
@@ -8,7 +8,7 @@ package com.liferay.jethr0.testsuite.repository;
 import com.liferay.jethr0.entity.repository.BaseEntityRepository;
 import com.liferay.jethr0.project.Project;
 import com.liferay.jethr0.testsuite.TestSuite;
-import com.liferay.jethr0.testsuite.dalo.ProjectsToTestSuitesDALO;
+import com.liferay.jethr0.testsuite.dalo.ProjectsToTestSuitesEntityRelationshipDALO;
 import com.liferay.jethr0.testsuite.dalo.TestSuiteDALO;
 
 import java.util.HashSet;
@@ -26,8 +26,9 @@ public class TestSuiteRepository extends BaseEntityRepository<TestSuite> {
 	public Set<TestSuite> getAll(Project project) {
 		Set<TestSuite> projectTestSuites = new HashSet<>();
 
-		Set<Long> testSuiteIds = _projectsToTestSuitesDALO.getChildEntityIds(
-			project);
+		Set<Long> testSuiteIds =
+			_projectsToTestSuitesEntityRelationshipDALO.getChildEntityIds(
+				project);
 
 		for (TestSuite testSuite : getAll()) {
 			if (!testSuiteIds.contains(testSuite.getId())) {
@@ -50,7 +51,8 @@ public class TestSuiteRepository extends BaseEntityRepository<TestSuite> {
 	}
 
 	@Autowired
-	private ProjectsToTestSuitesDALO _projectsToTestSuitesDALO;
+	private ProjectsToTestSuitesEntityRelationshipDALO
+		_projectsToTestSuitesEntityRelationshipDALO;
 
 	@Autowired
 	private TestSuiteDALO _testSuiteDALO;
